pyproject.toml
[project]
name = "repl-nix-asyncfastapichain"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.1.3",
    "asyncio>=3.4.3",
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.0",
    "chainlit>=1.2.0",
    "django-routers>=0.2",
    "email-validator>=2.2.0",
    "fastapi>=0.112.4",
    "jinja2>=3.1.4",
    "jose>=1.0.0",
    "langchain-community>=0.3.3",
    "langchain-experimental>=0.3.2",
    "langchain-openai>=0.2.3",
    "langchain-qdrant>=0.1.4",
    "langchain-text-splitters>=0.3.0",
    "langchain>=0.3.4",
    "langsmith>=0.1.136",
    "openai>=1.52.0",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.6.0",
    "pydantic>=2.9.2",
    "pymupdf>=1.24.12",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.9",
    "qdrant-client>=1.12.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.36",
    "sqlmodel>=0.0.22",
    "starlette>=0.37.2",
    "tavily-python>=0.5.0",
    "tenacity>=9.0.0",
    "tiktoken>=0.8.0",
    "uvicorn>=0.25.0",
    "wikipedia>=1.4.0",
]
